Instead of a division and modulo operation, match exhaustively
When computing the checksum, every other digit should be multiplied by two. If
the result is 10 or above, each digit should be added to the checksum
individually. We know that the original number is a single digit, so the results
that are 10 or above are 10, 12, 14, 16 and 18. Match them exhaustively instead
of doing one division and one modulo operation.