Add benchmark to measure speed of validate_isin()
The previous commit changes the digits from a dynamically allocated
vector to a fixed-size array. To demonstrate that this is a
performance-wise better approach, implement a benchmark of
validate_isin().

With Vec, the benchmark showed ~200ms/iter, with the static array this
is brought down to ~69ms/iter, more than a 50% improvement.