Unroll for-loop over digits to get rid of flag toggling
The flag deciding whether the odd or even numbers should be multiplied
by two is toggled while iterating over the digits.

Instead, split the iteration into two loops - one that handles the
digits that should just be added and one that multiplies the correct
groups. This only requires querying the flag twice and not l times.

The benchmark shows a gain of a few ms, from ~69 to about ~66 ms/iter.
It is a tiny improvement.