Rearrange input bytes reading for more performance
Instead of converting each byte to a number and then querying its value,
query the byte value directly. If the byte value is less than 58 we are
dealing with a single digit and only need one slot in the digits array,
otherwise we are dealing with a letter that requires two slots.

This brings the benchmark down to ~60ns/iter.